Owners cann book and sitters can view
Got it so that owners can view different sitters and then fill out a form to book that sitter, Sitters can view their jobs. need to fix so that form saves jobs to database.
