+++ petsitting/src/main/resources/templates/booking.html
@@ -42,10 +42,9 @@
 				<td th:text="${sitter.biography}" />
 				<td th:text="${sitter.averageRating}" />
 				
-
-				<td><input name="submit" type="submit" value="Book" />
 			</tr>
 		</table>
+		<input name="submit" type="submit" value="Book Sitter" />
 	</form>
 	<a href="/index.html">Return Home</a>
