+++ petsitting/src/main/resources/templates/viewJobs.html
@@ -8,23 +8,21 @@
 	<form>
 		<table id="Jobs">
 			<tr>
-				<th>First Name:</th>
-				<th>Last Name:</th>
-				<th>Pet Needs:</th>
+				<th>Owner Id:</th>
+				<th>Sitter Id:</th>
+				<th>Pet Needs</th>
 				<th>Job Date:</th>
 				<th>Job Time:</th>
 				<th>Status:</th>
 			</tr>
 
-			<tr th:each="sitter : ${sitters}">
-
-				<td th:text="${owner.firstName}" />
-				<td th:text="${owner.lastName}" />
-				<td th:text="${owner.petNeeds}" />
+			<tr th:each="job : ${jobs}">
+				<td th:text="${job.ownerId}" />
+				<td th:text="${job.sitterId}" />
+				<td th:text="${job.feedbackToSitter}" />
 				<td th:text="${job.jDate}" />
 				<td th:text="${job.jTime}" />
 				<td th:href="sitterAccept" th:text="${job.Status}" />
-				<td th:text="${job.feedbackToSitter}" />
 			</tr>
 		</table>
 	</form>
